+++ src/Router.tsx
@@ -3,8 +3,8 @@ import { Routes, Route, BrowserRouter } from 'react-router-dom';
 import ROUTE_PATH from '@Constant/routePath';
 
 import Hd from '@Pages/Hd';
-import JY from '@Pages/JY';
 import Landing from '@Pages/Landing';
+import MilkCheck from '@Pages/MilkCheck';
 
 function Router() {
   return (
@@ -12,7 +12,7 @@ function Router() {
       <Routes>
         <Route path={ROUTE_PATH.landing} element={<Landing />} />
         <Route path={ROUTE_PATH.hd} element={<Hd />} />
-        <Route path={ROUTE_PATH.jy} element={<JY/>} />
+        <Route path={ROUTE_PATH.milkcheck} element={<MilkCheck/>} />
       </Routes>
     </BrowserRouter>
   );
